# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: mini-evvy API
  description: |
    Event seating, bookings, payments, attendance, and background jobs.

  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /events/{eventId}/finalize-seating:
    post:
      tags: [Jobs]
      summary: Enqueue finalize seating job
      description: |
        Owner/admin. Event must be in seating_phase `open`. Enqueues a job that assigns seats
        to unbooked guests with barcodes (no emails). When the job completes, phase becomes
        `preview`. Poll job status; when complete, review via seating-preview and approve or reject.
      operationId: finalizeSeating
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobIdEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/EventNotFound"
        "409":
          description: Finalize already in progress or seating not open
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /events/{eventId}/seating-preview:
    get:
      tags: [Jobs]
      summary: Seating preview report
      description: Org member. Available when seating_phase is `preview` or `approved`. Lists guest assignments.
      operationId: getSeatingPreview
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeatingPreviewListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/EventNotFound"
  /events/{eventId}/seating-approve:
    post:
      tags: [Jobs]
      summary: Approve seating preview
      description: Owner/admin. Phase must be `preview`. Enqueues send_invitation for each active booking; sets phase `approved`.
      operationId: approveSeating
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/EventNotFound"
  /events/{eventId}/seating-reject:
    post:
      tags: [Jobs]
      summary: Reject seating preview
      description: Owner/admin. Phase must be `preview` or `approved`. Cancels pending/paid invited bookings, frees seats, sets phase `open`.
      operationId: rejectSeating
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/EventNotFound"
...
  /bookings/{bookingId}/resend-invitation:
    post:
      tags: [Bookings]
      summary: Re-send invitation email
      description: Org member. Booking must not be cancelled; guest must have email.
      operationId: resendInvitation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/bookingId"
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobIdEnvelope"
        "400":
          description: Guest email missing or booking not resendable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/BookingNotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from login/register. Subject is user UUID.
  parameters:
...
    eventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    bookingId:
      name: bookingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    EventNotFound:
      description: Event not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    BookingNotFound:
      description: Booking not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: BOOKING_NOT_FOUND
        message:
          type: string
...
    JobIdResponse:
      type: object
      required: [job_id]
      properties:
        job_id:
          type: string
          format: uuid
...
    JobIdEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/JobIdResponse"
...
    StatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
    StatusEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/StatusResponse"
    SeatingPreviewRow:
      type: object
      properties:
        booking_id:
          type: string
          format: uuid
        guest_name:
          type: string
        guest_email:
          type: string
        seat_code:
          type: string
        category_name:
          type: string
        category_code:
          type: string
          nullable: true
        section:
          type: string
          nullable: true
    SeatingPreviewListEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/SeatingPreviewRow"